backend/ai_service_fastapi/requirements.txt
fastapi==0.115.0
uvicorn==0.35.0
python-multipart==0.0.9
torch==2.3.1+cpu
torchvision==0.18.1+cpu
pydicom==2.4.4
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.3.0
pymongo==4.7.2
